+++ Kernel/include/std_buffers.h
@@ -1,3 +1,5 @@
+#include <stdint.h>
+
 #define STD_OUT 1
 #define STD_IN 0
 #define STD_ERR 2
@@ -5,7 +7,7 @@
 void setUpBuffers();
 void clearBuffer(int index);
 int writeCharBuffer(int index, char character);
-int writeIntBuffer(int index, int value,int base);
+int writeIntBuffer(int index, uint64_t value,int base);
 int writeStrBuffer(int index, char * str);
 char readCharBuffer(int index);
 int readStrBuffer(int index, char * str,int size);
